+++ src/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
+import { ErrorBoundary } from './components/error-boundary';
 import { App } from './App';
-import { ErrorBoundary } from './components/errorBoundary';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <ErrorBoundary>
